Moved types to project root
