+++ graphql_extensions/types.py
@@ -1,10 +1,28 @@
+import datetime
+
 from django.core.exceptions import ValidationError
 from django.core.validators import validate_email
 
 from graphene.types.scalars import Scalar
 from graphql.language import ast
 
-from .. import exceptions
+from . import exceptions
+
+
+class Timestamp(Scalar):
+
+    @classmethod
+    def serialize(cls, value):
+        return int(value.timestamp())
+
+    @classmethod
+    def parse_literal(cls, node):
+        if isinstance(node, ast.IntValue):
+            return cls.parse_value(node.value)
+
+    @classmethod
+    def parse_value(cls, value):
+        return datetime.datetime.fromtimestamp(value)
 
 
 class Email(Scalar):
