--- graphql_extensions/types/__init__.py
@@ -1,11 +1,0 @@
-from .camel_json import CamelJSON
-from .choices import Choices
-from .email import Email
-from .timestamp import Timestamp
-
-__all__ = [
-    'CamelJSON',
-    'Choices',
-    'Email',
-    'Timestamp',
-]
